requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.32.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.9.0
azure-keyvault-keys==4.10.0
azure-keyvault-secrets==4.9.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
isodate==0.7.2
lxml==5.3.0
msal==1.31.1
msal-extensions==1.2.0
poetry-core==2.0.1
portalocker==2.10.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
